refactor(rust): extract profiles.rs with profile CRUD commands
